chore: Move Card component to UI directory
