--- packages/webapp-libs/webapp-core/src/components/cards/cardContent/cardContent.component.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const CardContent = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
-  ({ className, ...props }, ref) => <div ref={ref} className={cn('p-6 pt-0', className)} {...props} />
-);
-CardContent.displayName = 'CardContent';
-
-export { CardContent };
