--- packages/webapp-libs/webapp-core/src/components/cards/cardContent/index.ts
@@ -1,1 +1,0 @@
-export * from './cardContent.component';
