--- packages/webapp-libs/webapp-core/src/components/cards/cardDescription/cardDescription.component.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const CardDescription = React.forwardRef<HTMLParagraphElement, React.HTMLAttributes<HTMLParagraphElement>>(
-  ({ className, ...props }, ref) => (
-    <p ref={ref} className={cn('text-sm text-muted-foreground', className)} {...props} />
-  )
-);
-CardDescription.displayName = 'CardDescription';
-
-export { CardDescription };
