--- packages/webapp-libs/webapp-core/src/components/cards/cardDescription/index.ts
@@ -1,1 +1,0 @@
-export * from './cardDescription.component';
