--- packages/webapp-libs/webapp-core/src/components/cards/cardFooter/cardFooter.component.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const CardFooter = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
-  ({ className, ...props }, ref) => <div ref={ref} className={cn('flex items-center p-6 pt-0', className)} {...props} />
-);
-CardFooter.displayName = 'CardFooter';
-
-export { CardFooter };
