--- packages/webapp-libs/webapp-core/src/components/cards/cardFooter/index.ts
@@ -1,1 +1,0 @@
-export * from './cardFooter.component';
