--- packages/webapp-libs/webapp-core/src/components/cards/cardHeader/cardHeader.component.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const CardHeader = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
-  ({ className, ...props }, ref) => (
-    <div ref={ref} className={cn('flex flex-col space-y-1.5 p-6', className)} {...props} />
-  )
-);
-CardHeader.displayName = 'CardHeader';
-
-export { CardHeader };
