--- packages/webapp-libs/webapp-core/src/components/cards/cardHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './cardHeader.component';
