--- packages/webapp-libs/webapp-core/src/components/cards/cardTitle/index.ts
@@ -1,1 +1,0 @@
-export * from './cardTitle.component';
