--- packages/webapp-libs/webapp-core/src/components/cards/cards.component.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-
-import { cn } from '../../lib/utils';
-
-const Card = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(({ className, ...props }, ref) => (
-  <div ref={ref} className={cn('rounded-lg border bg-card text-card-foreground shadow-sm', className)} {...props} />
-));
-Card.displayName = 'Card';
-
-export { Card };
