--- packages/webapp-libs/webapp-core/src/components/cards/index.ts
@@ -1,6 +1,0 @@
-export * from './cards.component';
-export * from './cardContent';
-export * from './cardDescription';
-export * from './cardFooter';
-export * from './cardHeader';
-export * from './cardTitle';
