+++ packages/webapp-libs/webapp-finances/src/routes/editSubscription/subscriptionPlanItem/subscriptionPlanItem.component.tsx
@@ -1,6 +1,6 @@
 import { FragmentType, StripeSubscriptionQueryQuery, getFragmentData } from '@sb/webapp-api-client/graphql';
 import { Button } from '@sb/webapp-core/components/buttons';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@sb/webapp-core/components/cards';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@sb/webapp-core/components/ui/card';
 import { cn } from '@sb/webapp-core/lib/utils';
 import { FormattedMessage } from 'react-intl';
