+++ packages/webapp-libs/webapp-tenants/src/components/tenantInvitationForm/tenantInvitationForm.component.tsx
@@ -1,6 +1,5 @@
 import { TenantUserRole } from '@sb/webapp-api-client';
 import { Button } from '@sb/webapp-core/components/buttons';
-import { Card, CardContent, CardHeader, CardTitle } from '@sb/webapp-core/components/cards';
 import {
   Form,
   FormControl,
@@ -15,6 +14,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@sb/webapp-core/components/forms';
+import { Card, CardContent, CardHeader, CardTitle } from '@sb/webapp-core/components/ui/card';
 import { cn } from '@sb/webapp-core/lib/utils';
 import { FormattedMessage, useIntl } from 'react-intl';
