+++ packages/webapp/src/routes/admin/admin.component.tsx
@@ -1,5 +1,5 @@
-import { Card } from '@sb/webapp-core/components/cards';
 import { H2, H4, Small } from '@sb/webapp-core/components/typography';
+import { Card } from '@sb/webapp-core/components/ui/card';
 import { ShieldCheckIcon } from 'lucide-react';
 import { FormattedMessage } from 'react-intl';
